requirements.txt
albumentations==0.5.2
tensorflow-gpu==2.3.0
pycocotools==2.0.2
